+++ geopandas/explore.py
@@ -278,6 +278,7 @@ GON (((180.00000 -16.06713, 180.00000...
     try:
         import branca as bc
         import folium
+        import re
         import matplotlib
         import matplotlib.colors as colors
         import matplotlib.pyplot as plt
@@ -417,7 +418,6 @@ GON (((180.00000 -16.06713, 180.00000...
 
             # colormap exists in matplotlib
             if cmap in plt.colormaps():
-
                 color = np.apply_along_axis(
                     colors.to_hex,
                     1,
@@ -455,7 +455,6 @@ GON (((180.00000 -16.06713, 180.00000...
 
             # get bins
             if scheme is not None:
-
                 if classification_kwds is None:
                     classification_kwds = {}
                 if "k" not in classification_kwds:
@@ -471,7 +470,6 @@ GON (((180.00000 -16.06713, 180.00000...
                 )
 
             else:
-
                 bins = np.linspace(vmin, vmax, 257)[1:]
                 binning = classify(
                     np.asarray(gdf[column][~nan_idx]), "UserDefined", bins=bins
@@ -618,10 +616,21 @@ GON (((180.00000 -16.06713, 180.00000...
     else:
         tooltip = None
         popup = None
+    # escape the curly braces {{}} for jinja2 templates
+    feature_collection = gdf.__geo_interface__
+    for feature in feature_collection["features"]:
+        for k in feature["properties"]:
+            # escape the curly braces in values
+            if type(feature["properties"][k]) == str:
+                feature["properties"][k] = re.sub(
+                    r"\{{2,}",
+                    lambda x: "{% raw %}" + x.group(0) + "{% endraw %}",
+                    feature["properties"][k],
+                )
 
     # add dataframe to map
     folium.GeoJson(
-        gdf.__geo_interface__,
+        feature_collection,
         tooltip=tooltip,
         popup=popup,
         marker=marker,
@@ -644,7 +653,6 @@ GON (((180.00000 -16.06713, 180.00000...
 
             _categorical_legend(m, caption, categories, legend_colors)
         elif column is not None:
-
             cbar = legend_kwds.pop("colorbar", True)
             colormap_kwds = {}
             if "max_labels" in legend_kwds: