+++ geopandas/tests/test_overlay.py
@@ -8,7 +8,6 @@ from shapely.geometry import Point, Polygon, LineString, GeometryCollection, box
 
 import geopandas
 from geopandas import GeoDataFrame, GeoSeries, overlay, read_file
-from geopandas import _compat
 from geopandas._compat import PANDAS_GE_20
 
 from geopandas.testing import assert_geodataframe_equal, assert_geoseries_equal
@@ -199,10 +198,7 @@ def test_overlay_nybb(how):
 
     # first, check that all bounds and areas are approx equal
     # this is a very rough check for multipolygon equality
-    if not _compat.PANDAS_GE_11:
-        kwargs = dict(check_less_precise=True)
-    else:
-        kwargs = {}
+    kwargs = {}
     pd.testing.assert_series_equal(
         result.geometry.area, expected.geometry.area, **kwargs
     )
@@ -409,7 +405,6 @@ def test_correct_index(dfs):
 
 
 def test_warn_on_keep_geom_type(dfs):
-
     df1, df2 = dfs
     polys3 = GeoSeries(
         [
@@ -724,7 +719,6 @@ def test_overlap_make_valid(make_valid):
 
 
 def test_empty_overlay_return_non_duplicated_columns():
-
     nybb = geopandas.read_file(geopandas.datasets.get_path("nybb"))
     nybb2 = nybb.copy()
     nybb2.geometry = nybb2.translate(20000000)